+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `web.abort()` and `Job.abort()` methods allowing to abort running tasks without deleting them. If a task is aborted, it cannot be restarted later, a new one needs to be created and submitted.
 - `FastDispersionFitter` for fast fitting of material dispersion data.
 - `Simulation.monitors_data_size` mapping monitor name to its data size in bytes.
+- Source with arbitrary user-specified time dependence through `CustomSourceTime`.
 
 ### Changed
 - Add `width` and `height` options to `Simulation.plot_3d`.